Rename resource names, services they create and add subscriptions
